% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{Boland.leopards}
\alias{Boland.leopards}
\title{Data from 2010 camera-trap survey of leopards in the Boland, South
Africa.}
\format{An object containing list with the following two data objects.
 \describe{
   \item{\code{Boland.CH}:}{ an \code{\link{secr}} capture history object with camera-trap
   detections of leopards in the form of binary data over 13 occasions.}
   \item{\code{Boland.mask}:}{ an \code{\link{secr}} mask object for the data in \code{Boland.CH}.
   It contans the following covariates:
   \describe{
   \item{\code{alt}:}{altitude}
   \item{\code{xalt}:}{product of x and altitude, scaled by dividing by its mean}
   \item{\code{yalt}:}{product of y and altitude, scaled by dividing by its mean}
   \item{\code{Natural}:}{binary variable: 1=Natural land use category, 0=not}
   \item{\code{Cultivated}:}{binary variable: 1=Cultivated land use category, 0=not}
   \item{\code{Degraded}:}{binary variable: 1=Degraded land use category, 0=not}
   \item{\code{Plantation}:}{binary variable: 1=Plantation land use category, 0=not}
   \item{\code{dist2.Urban}:}{distance to closest Urban land use category cell}
   \item{\code{dist2.Water}:}{distance to closest Water land use category cell}
   \item{\code{dist2.Natural}:}{distance to closest Natural land use category cell}
   \item{\code{Landuse}:}{Landuse category (1=Natural, 2=Cultivated, 3=Degraded, 4=Plantation)}
   \item{\code{lufactor}:}{Landuse category as a factor variable}
   }
 }
}}
\source{
\code{Boland.CH} and \code{Boland.mask} were provided by the Cape Leopard Trust
Boland Leopard Project (http://capeleopard.org.za/research/leopard/boland).
}
\usage{
data(Boland.leopards)
}
\description{
The data are from a camera-trap survey of leopards in the Boland region of
South Africa. Cameras were checked weekly and data recorded in binary format (i.e. only
whether or not an animal was detected on each occasion, not the number of detections
within occasions).
}

