% Generated by roxygen2 (4.0.1): do not edit by hand
\name{D.bspline}
\alias{D.bspline}
\title{B-spline density surface evaluation.}
\usage{
D.bspline(Dbeta, mask, ngroup = 1, nsession = 1)
}
\arguments{
\item{Dbeta}{model parameter values.}

\item{mask}{secr mask object as output by prepare.mask.bases, i.e. with attr(mask,"bases")
added as appropriate.}

\item{ngroup}{see help for \code{userDfun} in package \code{\link{secr}}.}

\item{nsession}{number of sessions - see help for \code{userDfun} in package
\code{\link{secr}}.}
}
\description{
Calculates density at all points in mask, using basis functions in attr(mask,"bases")
and parameter values in Dbeta. Currently only works for 1 session and with ngroup=1.
}
\details{
This function conforms to the specifiction for user-defined functions given in the
vignette "modelling density surfaces" of package "secr". It is intended for use with
the function secr.fit() via the userDfn component of the details argument, as follows:
secr.fit(...,details = list(userDfn = userDfn0),...).
}

