% Generated by roxygen2 (4.0.1): do not edit by hand
\name{derived}
\alias{derived}
\title{Derived density calculation for secrgam objects}
\usage{
derived(object, ..., use.secr = FALSE)
}
\arguments{
\item{object}{a fitted secrgam model (returned from \code{\link{secrgam.fit}})}

\item{...}{additional arguments passed to \code{derived} for package \code{\link{secr}}.}

\item{use.secr}{logical which, when TRUE, forces estimation with \code{derived} from
package \code{\link{secr}}.}
}
\description{
If the density model is uniform, calls \code{derived} from package
\code{\link{secr}} to calculate density, else returns an error message.
}
\details{
The density for non-uniform density models is not yet implemented.
See \code{derived} in package \code{\link{secr}} for details of derived calculations
when density is uniform.
}

