% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitted.secrgam}
\alias{fitted.secrgam}
\title{Extract fitted density values from an secrgam model}
\usage{
\method{fitted}{secrgam}(fit, mask = NULL)
}
\arguments{
\item{fit}{a fitted \code{\link{secrgam}} or \code{\link{secr}} model}

\item{mask}{a \code{\link{mask}} object}
}
\value{
value...
}
\description{
description...
}
\details{
details...
}
\examples{
# construct a tensor product model to represent a bivariate normal
data(Boland.leopards)
fitted(Boland.fit)
}

