% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.density.design.matrix}
\alias{make.density.design.matrix}
\title{Construct a design matrix from a density model}
\usage{
make.density.design.matrix(Dmodel, mask)
}
\arguments{
\item{model}{density model}

\item{mask}{secr mask object.}
}
\value{
Returns a matrix with number of rows equal to the number of rows in \code{mask}, and number of columns equal to the number of parameters required to implement the regresssion spline specified in \code{model}.
}
\description{
Helper function to create a design matrix from a conventional density formula (e.g. including \code{\link{s}} and/or \code{\link{te}} terms) and a mask object (i.e. a dataframe of covariates).
}
\details{
Uses the \code{\link{gam}} function from the \code{\link{mgcv}} package to generate the design matrix (by setting \code{fix = FALSE}).
}
\examples{
data(Boland.leopards1)
X = make.density.design.matrix(~s(x, k = 3), Boland.mask1)
head(X)
}

