% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.secrgam}
\alias{plot.secrgam}
\title{Plot fitted density surface or detection function}
\usage{
\method{plot}{secrgam}(fit, type = c("default", "density", "smooth"),
  mask = NULL, ...)
}
\arguments{
\item{fit}{a fitted \code{\link{secrgam}} or \code{\link{secr}} model}

\item{type}{the type of plot to produce (see Details)}

\item{mask}{an optional \code{\link{mask}} object to use when \code{type = "density"}}

\item{...}{aditional arguments to pass to the relevant ploting function}
}
\description{
Plot the fitted density surface or fitted detection function from an \code{secrgam} model.
}
\details{
\tabular{ll}{
type      \tab plotting function \cr
"default" \tab \code{\link{plot.secr}} \cr
"density" \tab \code{\link{prep4image}} \cr
"smooth"  \tab \code{\link{plotDgam}} \cr

}
}
\examples{
data(Boland.leopards1)
data(Boland.fits1)
op = par(no.readonly = TRUE)
par(mfrow = c(1,2), oma = c(0,0,0,0), mar = c(4,4,4,4))

# density surface
plot(fit1.a3.dW3,type="density",asp=1,key=FALSE,col=tim.colors(30))
# plot detection function
plot(fit1.a3.dW3,xval=seq(0,11000,length=100), main = "Density")

# plot smooths
plot(fit1.a3.dW3,type="smooth")

par(op)
}
\seealso{
\code{\link{fitted.secrgam}}
}

