% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotDgam_beta}
\alias{plotDgam_beta}
\title{SECR Density spline plotting.}
\usage{
plotDgam_beta(fit, type = "link", mask.rug = FALSE, det.rug = TRUE,
  npts = 200, show.knots = FALSE)
}
\arguments{
\item{fit}{object returned by \code{\link{secrgam.fit}}.}

\item{type}{either \code{"link"} (for plot on link funtion scale) or
\code{"response"} (for plot on response scale).}

\item{mask.rug}{\code{TRUE} if a rugplot of locations of mask points in the
covariate dimension is to be plotted.}

\item{det.rug}{\code{TRUE} if a rugplot of locations of detectors in the
covariate dimension is to be plotted.}

\item{npts}{number of points on x-axis at which to evaluate smooth.}

\item{show.knots}{if \code{TRUE}, puts crosses on smooth at knot locations.
  NOT YET IMPLEMENTED.}
}
\description{
Plots regression spline smooth components of an secr density
model.
}
\details{
(None as yet)
}
\examples{
op = par(no.readonly = TRUE)

data(Boland.fits1)
plotDgam(fit1.a3)

par(mfrow = c(1,2))
data(Boland.fits2)
plotDgam(fit2.a3.dW3)

par(op)
}

