% Generated by roxygen2 (4.0.1): do not edit by hand
\name{trap.covar}
\alias{trap.covar}
\title{Returns covariate values of mask point closest to each trap.}
\usage{
trap.covar(traps, mask, cname)
}
\arguments{
\item{traps}{Trap object.}

\item{mask}{Mask object.}

\item{cname}{Name of the covariate associated with the mask that you want to attach to the traps.}
}
\value{
Returns a vector with covariate values associated with each trap, in the same order as they appear
in \code{traps}
}
\description{
Finds mask point closest to each trap then assigns to the trap the covariate value in
\code{attributes(mask)$covariates[[cname]]} that is associated with this mask point.
}
\details{
Finds mask point closest to each trap then assigns to the trap the covariate value in
attributes(mask)$covariates[[cname]] that is associated with this mask point.
}

