\docType{data}
\name{Boland.alt.image}
\alias{Boland.alt.image}
\title{Altitudes in the Boland (in fromat for plotting).}
\format{A list with the following elements: $x (vector of unique longitudes, in increasing order),
$y (vector of unique longitudes, in increasing order) and $z (matrix of altitudes)..}
\source{
www.ngdc.noaa.gov/mgg/global/
}
\usage{
data(Boland.alt.image)
}
\description{
Altitudes of terrain in the the Boland, South Africa.
}
\references{
Amante, C. and B. W. Eakins, ETOPO1 1 Arc-Minute Global
Relief Model: Procedures, Data Sources and Analysis. NOAA
Technical Memorandum NESDIS NGDC-24, 19 pp, March 2009.
}

