\docType{data}
\name{Boland.landuse.image}
\alias{Boland.landuse.image}
\title{Land use class in the Boland (in fromat for plotting).}
\format{A list with the following elements: $x (vector of unique longitudes, in increasing order),
$y (vector of unique longitudes, in increasing order) and $z (matrix of land use
categories: 1="Natural", 2="Cultivated", 3="Degraded", 4="Urban", 5="Water",
6="Plantation").}
\source{
http://bgis.sanbi.org/landcover/project.asp
}
\usage{
data(Boland.landuse.image)
}
\description{
Land use class of terrain in the the Boland, South Africa.
}
\references{
http://bgis.sanbi.org/landcover/Landcover2009.pdf
}

