\docType{data}
\name{Boland.landuse}
\alias{Boland.landuse}
\title{Land use class in the Boland.}
\format{A data frame with the following elements $x, $y and $z, where $x and $y are longitude
and latitude and $z is land use category: 1="Natural", 2="Cultivated", 3="Degraded", 4="Urban",
5="Water", 6="Plantation".}
\source{
http://bgis.sanbi.org/landcover/project.asp
}
\usage{
data(Boland.landuse)
}
\description{
Land use class of terrain in the Boland, South Africa.
}
\references{
http://bgis.sanbi.org/landcover/Landcover2009.pdf
}

