% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fitted.secrgam}
\alias{fitted.secrgam}
\title{Get fitted density values from an secrgam model}
\usage{
\method{fitted}{secrgam}(fit, mask = NULL, type = c("individual", "mean"))
}
\arguments{
\item{fit}{a fitted \code{\link{secrgam}} or \code{\link{secr}} model}

\item{mask}{a \code{\link{mask}} object}

\item{type}{the type of fitted value required (use \code{type = "individual"} for fitted values at each individual mask point, or \code{type = "mean"} for the estmated mean density across all mask points).}
}
\value{
If \code{se = FALSE} a numeric vector of fitted values is returned (or a single value if \code{type = "mean"}). If \code{se = TRUE} a list is returned with elements \code{lower}, \code{estimate} and \code{upper}.
}
\description{
Get fitted density values for each mak point or average density across the mask.
}
\examples{
data(Boland.leopards1)
data(Boland.fits1)
op = par(no.readonly = TRUE)

# fitted density values at each mask point
Dhat = fitted(fit1.a3)
par(mfrow = c(1,1))
prep4image(cbind(x = fit1.a3$mask$x,
                 y = fit1.a3$mask$y,
                 z = Dhat), asp = 1)

# fitted mean density across all mask points
Dhat = fitted(fit1.a3, type = "mean") ; Dhat
M = nrow(fit1.a3$mask) ; M # number of mask points
a = attr(fit1.a3$mask, "a") ; a # area of each mask cell
A = M * a ; A # total area of mask
Nhat = Dhat * A ; Nhat # abundance within mask
}
\seealso{
\link{plot.secrgam}
}

