% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.density.design.matrix}
\alias{make.density.design.matrix}
\title{Construct a design matrix from a density model}
\usage{
make.density.design.matrix(Dmodel, mask)
}
\arguments{
\item{model}{density model}

\item{mask}{secr mask object.}
}
\value{
Returns a matrix with \code{n} rows and \code{p} columns.
}
\description{
description...
}
\examples{
data(Boland.leopards1)
X = make.density.design.matrix(~s(x, k = 3), Boland.mask1)
head(X)
}

