\name{make.density.design.matrix}
\alias{make.density.design.matrix}
\title{Construct a design matrix from a density model}
\usage{
make.density.design.matrix(Dmodel, mask)
}
\arguments{
  \item{model}{secr model specification list.}

  \item{mask}{secr mask object.}
}
\value{
return...
}
\description{
description...
}
\details{
details...
}

