% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.secrgam}
\alias{plot.secrgam}
\title{Plot fitted density surface or detection function}
\usage{
\method{plot}{secrgam}(fit, density = TRUE, mask = NULL, ...)
}
\arguments{
\item{fit}{a fitted \code{\link{secrgam}} or \code{\link{secr}} model}

\item{density}{if \code{TRUE} then the fitted density surface will be plotted, if \code{FALSE} then \code{\link{plot.secr}} is used to plot the fitted detection function}

\item{mask}{an optional \code{\link{mask}} object to use for plotting the fitted density surface}

\item{...}{aditional arguments to pass to \code{prep4image} (if \code{density = TRUE}), or \code{plot.secr} (if \code{density = FALSE})}
}
\description{
Plot the fitted density surface or fitted detection function from an \code{secrgam} model.
}
\examples{
data(Boland.leopards1)
data(Boland.fits1)
op = par(no.readonly = TRUE)
par(mfrow = c(1,2), oma = c(0,0,0,0), mar = c(4,4,4,4))

# density surface
plot(fit1.a3, asp = 1, key = FALSE, main = "Density")

# detection function
plot(fit1.a3, density = FALSE, xval = 0:20000, main = "Detection function")

par(op)
}
\seealso{
\code{\link{fitted.secrgam}}
}

