% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotDgam}
\alias{plotDgam}
\title{SECR Density spline plotting.}
\usage{
plotDgam(fit, scale = "response", mask.rug = FALSE, det.rug = TRUE,
  bounds = TRUE, npts = 200, main = TRUE)
}
\arguments{
\item{fit}{object returned by \code{\link{secrgam.fit}}.}

\item{scale}{either \code{"link"} (for plot on link funtion scale) or
\code{"response"} (for plot on response scale).}

\item{mask.rug}{\code{TRUE} if a rugplot of locations of mask points in the
covariate dimension is to be plotted.}

\item{det.rug}{\code{TRUE} if a rugplot of locations of detectors in the
covariate dimension is to be plotted.}

\item{bounds}{\code{TRUE} if 95\% confidence bounds of the smooth are to be
plotted.}

\item{npts}{number of points on x-axis at which to evaluate smooth.}

\item{main}{if TRUE, a main heading reflecting the smooth term in the model
that is being plotted, is added to the plot; else no main heading is plotted.}
}
\description{
Plots regression spline smooth components of an secr density
model.
}
\details{
Plots smooth on link or response scale, together with rug plot showing locations
of detectors in covariate space (if det.rug==TRUE), rug plot showing locations
of mask points (if mask.rug==TRUE) and 95\% confidence bounds (if bounds=TRUE).

This function is not currently able to deal with interactions between sessions
and other covariates. It can plot a smooth of session or of mask covariates
but in the case of mult-session fits, it will plot only smooths of non-session
covariates for the first session, using the firs session mask.

Be aware that using mask.rug=TRUE will result in it taking quite a while to
produce the plot
}
\examples{
op = par(no.readonly = TRUE)

data(Boland.fits1)
plotDgam(fit1.a3)

par(mfrow = c(1,2))
data(Boland.fits2)
plotDgam(fit2.a3.dW3)

par(op)
}

