% Generated by roxygen2 (4.0.1): do not edit by hand
\name{prep4image}
\alias{prep4image}
\title{Prepares data frame for plotting with image/contour/persp.}
\usage{
prep4image(data, plot = TRUE, contour = TRUE, key = TRUE, ...)
}
\arguments{
\item{data}{a data frame with columns x and y being Cartesian coordinates,
and z being the values of some variable at each coordinate.}

\item{plot}{if \code{TRUE} then an image plot will be drawn using
\code{\link{image.plot}}}

\item{contour}{if \code{TRUE} then contours will be added if (only used when
\code{plot=TRUE})}

\item{key}{logical for whether or not to plot color key with plot}

\item{...}{other arguments to pass to \code{\link{image.plot}} (only used
  when \code{plot=TRUE})}
}
\description{
From an input data frame with columns x, y and z, this function
  creates a list with elements x, y and z in a format suitable for passing to
  functions \code{\link{image}}, \code{\link{contour}} or
  \code{\link{persp}}. The coordinates in \code{data} are assumed to come
  from a 2D grid of points.
}
\details{
Sorts z on values of x first, then y, then creates a matrix of
  z-values from this. Returns a list with elements x (unique values of x, in
  increasing order), y (unique values of y, in increasing order) and z
  (matrix of z-values in appropriate order for image/contour/persp).
}

