\docType{data}
\name{WCape.alt}
\alias{WCape.alt}
\title{Altitudes in Western Cape.}
\format{A data frame with the following elements $x, $y and $z, where $x and $y are longitude
and latitude and $z is altitude (in m).}
\source{
www.ngdc.noaa.gov/mgg/global/
}
\usage{
data(WCape.alt)
}
\description{
Altitudes of terrain in the Western Cape, South Africa.
}
\references{
Amante, C. and B. W. Eakins, ETOPO1 1 Arc-Minute Global
Relief Model: Procedures, Data Sources and Analysis. NOAA
Technical Memorandum NESDIS NGDC-24, 19 pp, March 2009.
}

